+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
+## [Unreleased]
+### Fixed
+- Ensure that the `TxnsController` inherits from this engine's `ApplicationController`
+
 ## [0.4.0] - 2016-08-08
 ### Changed
 - Change migration name to include "bloom_remit" to avoid conflicts